+++ src/App.js
@@ -1,38 +1,27 @@
 import React from "react";
-import Directory from "./Components/directory/directory.component";
+// import Home from "./Components/routes/home/home.component";
+import Home from "./routes/home/home.component";
+// import Navigation from "./Components/routes/navigation/navigation.component";
+import Navigation from "./routes/navigation/navigation.component";
+import { Route, Routes } from "react-router-dom";
 
-const App = () => {
-  const categories = [
-    {
-      id: 1,
-      title: "Hats",
-      imageUrl: "https://i.ibb.co/cvpntL1/hats.png",
-    },
-    {
-      id: 2,
-      title: "Jackets",
-      imageUrl: "https://i.ibb.co/px2tCc3/jackets.png",
-    },
-    {
-      id: 3,
-      title: "Sneakers",
-      imageUrl: "https://i.ibb.co/0jqHpnp/sneakers.png",
-    },
-    {
-      id: 4,
-      title: "Womens",
-      imageUrl: "https://i.ibb.co/GCCdy8t/womens.png",
-    },
-    {
-      id: 5,
-      title: "Mens",
-      imageUrl: "https://i.ibb.co/R70vBrQ/men.png",
-    },
-  ];
-
+const Shop = () => {
   return (
     <>
-      <Directory categories={categories} />
+      <h1>This is a shop page</h1>
+    </>
+  );
+};
+
+const App = () => {
+  return (
+    <>
+      <Routes>
+        <Route path="/" element={<Navigation />}>
+          <Route index={true} element={<Home />} />
+          <Route path="shop" element={<Shop />} />
+        </Route>
+      </Routes>
     </>
   );
 };
