+++ src/Components/directory/directory.component.jsx
@@ -13,4 +13,5 @@ const Directory = ({ categories }) => {
     </>
   );
 };
+
 export default Directory;
